Support autofixing ImplicitAsioJoin for instance method calls (#36)
Currently, the ImplicitAsioJoin autofix does not behave correctly
for instance method invocations, as it has no knowledge of the
variable that the instance method was invoked on and thus always
ends up outputting a static method call.

As a fix, pass the name of the variable that the instance method
is invoked on to check_implicit_asio_join() if appropriate,
and have get_async_version_name() output an instance method invocation
on this variable if provided and in an autofix context. Add tests.