ImplicitAsioJoin: Improve formatting for long method chains
Currently, the ImplicitAsioJoin autofix leaves the LHS expression
completely untouched when asyncifying instance method calls.
This is suboptimal if the LHS expression is a long method chain
that is broken into multiple lines, as the boundaries of the resulting
expression will be difficult to discern, e.g.:

```hack
async function test(A $bar): Awaitable<void> {
	$foo = (await $bar->getSomething()
		->call()
		->otherCall())
		->sync();
}
```

So, add a new replacement type that removes newlines and leading whitespace
within a target source range and apply it to the LHS expression (if one exists)
as part of the ImplicitAsioJoin autofix.