Add ClassnameUsedAsString issue with autofix in argument context
I ran an updated Hack typechecker against a webapp checkout in a local VM.
One of the noisiest new issues is the use of a classname literal in a string context;
Hack now wants us to explicitly use `nameof C` instead of `C::class` here to avoid
an implicit conversion to string at runtime.

So, add a new issue type for detecting classname<T> used in a string context,
and have it detect and autofix the vastly most common case of using a classname
literal as an argument to a function which takes string. The Hack typechecker
does not seem to raise an error as of yet if the classname is passed to the function
as a variable rather than a literal and most webapp cases seem to directly pass
a literal anyways, so ignore this case in Hakana as well.
We can always make our issue more comprehensive as a followup.