ClassnameUsedAsString: Improve classname resolution (#48)
In https://github.com/slackhq/hakana/pull/45, @muglug suggested we implement
a better classname resolution mechanism for the ClassnameUsedAsString autofix.
I did not see one at the time, but it turns out there's a simple one: we can
take the class constant expression that forms `C::class` and simply get the literal
class name string from it, which is what we already do when determining the
literal string value of a `nameof` expression during expression analysis.

The main benefit of this is that not only does this automatically return
a correctly qualified class name string, it also correctly handles references
like self/static/parent, since it just returns the literal. The latter currently
get resolved by the autofix, so e.g. static::class would end up being autofixed
to whatever `static` was referring to, which is undesired.

Add tests for the behavior.