Exempt variables holding a `classname` from `ClassnameUsedAsString`
The Hack typechecker requires the RHS argument of `nameof` to be a classname literal,
such as `C` or `self`; `nameof $some_var_holding_classname` does not typecheck.
So, keep variables holding a classname untouched even when they are used in a string context,
as they do not cause type errors anyways.