Prepare for Rust 2024's `if let` rescoping (#54)
Rust 2024 changes the lifetime of temporary values in the scrutinee
of `if let` blocks.[1] The autofix that runs as part of `cargo fix`
is conservative and will mangle any affected `if let` blocks into equivalent
`match` expressions.

Most of the affected blocks in Hakana can actually be better expressed by
chaining `.map_or` or `.unwrap_or_else`, which is less verbose.
Update these accordingly.

[1] https://doc.rust-lang.org/nightly/edition-guide/rust-2024/temporary-if-let-scope.html