ClassnameUsedAsString: Handle keyset values and dict keys (#67)
The updated Hack typechecker complains when a classname literal
in the form of `C::class` is used as a dict key or keyset value.
Handle these akin to classname function arguments used in a string context.
This slightly changes the reported offsets of the latter as we now
report the exact position of the offending argument rather than
the position of the parent function call.