Don't fake calling_class during top-level constant analysis
We currently set a fake `calling_class` when analyzing a top-level constant.
This causes issues if any of the analysis code then calls `.is_production`,
which expects it to be a valid classlike ID if it's set.

Removing this shouldn't impact issue caching (AFAIK), since issue symbols
are derived from calling_functionlike_id, which we do not set.