Ensure get_namespace() returns a consistent namespace (#76)
The StatementsAnalyzer instance used to analyze class constant definitions
uses a different FileAnalyzer instance that is not updated to reflect
potentially enclosing namespace definitions. This didn't seem to cause
issues before because nothing was calling ScopeAnalyzer::get_namespace().

Fix it by creating a new StatementsAnalyzer for each top-level def
so that we can pass the same FileAnalyzer ref with a correct namespace.