Use existing name resolution algorithm for `nameof`
The current class name resolution mechanism for `nameof` does not handle
names aliased via `use` statements correctly. However, it turns out we can
replace the bespoke logic with `get_id_name()` instead which already takes care
of this. In doing so, remove the `NameofUsedOutsideClassWithoutLiteral` issue
as it's somewhat exotic and there's little value in keeping around handling for it
now that we're using an existing lookup implementation.