Raise error for the error silencing operator (#80)
Upcoming HHVM drops runtime support for the `@` error silencing
operator. oxidized doesn't actually output expressions for `@`
in typechecker mode, only in codegen mode which doesn't work
with Hakana, so configure the parser to raise errors when `@`
is used and catch the resulting lowering error. Update other
tests that changed now that we check lowering errors.